feat(nes): iNES 2.0 header support (#271) (#288)
NES 2.0 extends the 16-byte iNES header with a 12-bit mapper
number, sub-mapper, extended PRG/CHR sizes, PRG-RAM/EEPROM/CHR-RAM
sizes, and TV-system hints. Modern repacks increasingly ship v2.0
headers; the v1.0 parser was misreading those bytes as garbage in
flag8-12.

internal/nes/ines.go:
- ROM.Mapper widened to uint16 (12-bit per spec; low byte still
  populated from flag6/7 for v1.0 compat).
- New fields: SubMapper, PRGRAMSize, EEPROMSize, CHRRAMSize,
  TVSystem.
- TVSystem enum (NTSC / PAL / Dual / Dendy) with String for
  diagnostics.
- parseNES2Extensions reads flag8-12 + writes the derived fields
  when the v2.0 ID bits (flag7 0b00001000) are set.
- extendedBankCount handles both linear (high nibble < $F) +
  exponential (high == $F → M × 2^E) PRG/CHR sizing.

cart.Open still dispatches on Mapper's low byte. Mapper-number-
above-255 ROMs aren't supported yet — they error out via the
existing "unsupported mapper" path. Will be filled as VRC /
FME-7 mappers (issues #269 / #270) land.

Tests (ines_v2_test.go):
- v2.0 ID bit detection.
- 12-bit mapper assembly from flag6/7/8 nibbles.
- Sub-mapper extraction.
- PRG-RAM + EEPROM + CHR-RAM size decoding.
- TV-system enum decoding for all four cases.
- Legacy v1.0 header stays v1.0 (NES2=false, extensions zeroed).

Regression: full test sweep + lint green. mapper_test cases
updated to use uint16 want values.

Closes #271.

+++ internal/nes/ines_test.go
@@ -75,7 +75,7 @@ func TestParse_MapperByteAssembly(t *testing.T) {
 	cases := []struct {
 		name         string
 		flag6, flag7 byte
-		want         uint8
+		want         uint16
 	}{
 		{"mapper 0 (NROM)", 0x00, 0x00, 0},
 		{"mapper 1 (MMC1)", 0x10, 0x00, 1},
